+++ internal/adapters/httpserver/handlers/response.go
