+++ internal/adapters/httpserver/handlers/scrape.go
