+++ internal/adapters/httpserver/handlers/search.go
